+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # 5.2.11
-## XX/XX/XXXX
+## XX/XX/2016
 
 1. [Common](#common)
     1. [](#new)
         - Added `very-large-desktop-range` to `breakpoint` mixin in order to be used when working with screen resolutions of 1920px+
+    3. [](#bugfix)
+        - System outlines should not be able to assign to pages (Fixes #1146)
 2. [Joomla](#joomla)
     3. [](#bugfix)
         - Admin: Fix potential fatal error when saving Outline Assignments